get rid of the extra file base_dimension_type and move that to dimension_type
